Retry transient http error codes when checking GCS
PiperOrigin-RevId: 321202206
