+++ dsub/providers/google_base.py
@@ -24,20 +24,16 @@ import io
 import json
 import os
 import re
-import socket
-import ssl
-import sys
 import warnings
 
 from .._dsub_version import DSUB_VERSION
 import googleapiclient.discovery
 import googleapiclient.errors
-from httplib2 import ServerNotFoundError
 from ..lib import job_model
+from ..lib import retry_util
 import pytz
 import retrying
 from six.moves import range
-import six.moves.http_client
 
 import google.auth
 from google.oauth2 import service_account
@@ -62,22 +58,6 @@ DEFAULT_SCOPES = [
     'https://www.googleapis.com/auth/monitoring.write',
 ]
 
-
-# Transient errors for the Google APIs should not cause them to fail.
-# There are a set of HTTP and socket errors which we automatically retry.
-#  429: too frequent polling
-#  50x: backend error
-TRANSIENT_HTTP_ERROR_CODES = frozenset([429, 500, 503, 504])
-
-# Auth errors should be permanent errors that a user needs to go fix
-# (enable a service, grant access on a resource).
-# However we have seen them occur transiently, so let's retry them when we
-# see them, but not as patiently.
-HTTP_AUTH_ERROR_CODES = frozenset([401, 403])
-
-# Socket error 32 (broken pipe) and 104 (connection reset) should
-# also be retried
-TRANSIENT_SOCKET_ERROR_CODES = frozenset([32, 104])
 
 # When attempting to cancel an operation that is already completed
 # (succeeded, failed, or canceled), the response will include:
@@ -159,11 +139,6 @@ _ZONES = [
     'us-west2-b',
     'us-west2-c',
 ]
-
-
-def _print_error(msg):
-  """Utility routine to emit messages to stderr."""
-  print(msg, file=sys.stderr)
 
 
 def get_zones(input_list):
@@ -481,118 +456,18 @@ def cancel(batch_fn, cancel_fn, ops):
   return canceled_ops, error_messages
 
 
-def _print_retry_error(exception, verbose):
-  """Prints an error message if appropriate."""
-  if not verbose:
-    return
-
-  now = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S.%f')
-  try:
-    status_code = exception.resp.status
-  except AttributeError:
-    status_code = ''
-
-  _print_error('{}: Caught exception {} {}'.format(now,
-                                                   type(exception).__name__,
-                                                   status_code))
-  _print_error('{}: This request is being retried'.format(now))
-
-
-def retry_api_check(exception, verbose):
-  """Return True if we should retry. False otherwise.
-
-  Args:
-    exception: An exception to test for transience.
-    verbose: If true, output retry messages
-
-  Returns:
-    True if we should retry. False otherwise.
-  """
-  if isinstance(exception, googleapiclient.errors.HttpError):
-    if exception.resp.status in TRANSIENT_HTTP_ERROR_CODES:
-      _print_retry_error(exception, verbose)
-      return True
-
-  if isinstance(exception, socket.error):
-    if exception.errno in TRANSIENT_SOCKET_ERROR_CODES:
-      _print_retry_error(exception, verbose)
-      return True
-
-  if isinstance(exception, google.auth.exceptions.RefreshError):
-    _print_retry_error(exception, verbose)
-    return True
-
-  # For a given installation, this could be a permanent error, but has only
-  # been observed as transient.
-  if isinstance(exception, ssl.SSLError):
-    _print_retry_error(exception, verbose)
-    return True
-
-  # This has been observed as a transient error:
-  #   ServerNotFoundError: Unable to find the server at genomics.googleapis.com
-  if isinstance(exception, ServerNotFoundError):
-    _print_retry_error(exception, verbose)
-    return True
-
-  # Observed to be thrown transiently from auth libraries which use httplib2
-  # Use the one from six because httlib no longer exists in Python3
-  # https://docs.python.org/2/library/httplib.html
-  if isinstance(exception, six.moves.http_client.ResponseNotReady):
-    _print_retry_error(exception, verbose)
-    return True
-
-  return False
-
-
-def retry_api_check_quiet(exception):
-  return retry_api_check(exception, False)
-
-
-def retry_api_check_verbose(exception):
-  return retry_api_check(exception, True)
-
-
-def retry_auth_check(exception, verbose):
-  """Specific check for auth error codes.
-
-  Return True if we should retry.
-
-  False otherwise.
-  Args:
-    exception: An exception to test for transience.
-    verbose: If true, output retry messages
-
-  Returns:
-    True if we should retry. False otherwise.
-  """
-  if isinstance(exception, googleapiclient.errors.HttpError):
-    if exception.resp.status in HTTP_AUTH_ERROR_CODES:
-      _print_retry_error(exception, verbose)
-      return True
-
-  return False
-
-
-def retry_auth_check_quiet(exception):
-  return retry_auth_check(exception, False)
-
-
-def retry_auth_check_verbose(exception):
-  return retry_auth_check(exception, True)
-
-
 # Exponential backoff retrying API discovery.
 # Maximum 23 retries.  Wait 1, 2, 4 ... 64, 64, 64... seconds.
 @retrying.retry(
     stop_max_attempt_number=24,
-    retry_on_exception=retry_api_check_verbose,
+    retry_on_exception=retry_util.retry_api_check_verbose,
     wait_exponential_multiplier=500,
     wait_exponential_max=64000)
 # For API errors dealing with auth, we want to retry, but not as often
 # Maximum 4 retries. Wait 1, 2, 4, 8 seconds.
 @retrying.retry(
     stop_max_attempt_number=5,
-    retry_on_exception=retry_auth_check_verbose,
+    retry_on_exception=retry_util.retry_auth_check_verbose,
     wait_exponential_multiplier=500,
     wait_exponential_max=8000)
 def setup_service(api_name, api_version, credentials=None):
@@ -645,14 +520,14 @@ class Api(object):
   # Maximum 23 retries.  Wait 1, 2, 4 ... 64, 64, 64... seconds.
   @retrying.retry(
       stop_max_attempt_number=24,
-      retry_on_exception=retry_api_check_verbose,
+      retry_on_exception=retry_util.retry_api_check_verbose,
       wait_exponential_multiplier=500,
       wait_exponential_max=64000)
   # For API errors dealing with auth, we want to retry, but not as often
   # Maximum 4 retries. Wait 1, 2, 4, 8 seconds.
   @retrying.retry(
       stop_max_attempt_number=5,
-      retry_on_exception=retry_auth_check_verbose,
+      retry_on_exception=retry_util.retry_auth_check_verbose,
       wait_exponential_multiplier=500,
       wait_exponential_max=8000)
   def execute_verbose(self, api):
@@ -670,14 +545,14 @@ class Api(object):
   # Maximum 23 retries.  Wait 1, 2, 4 ... 64, 64, 64... seconds.
   @retrying.retry(
       stop_max_attempt_number=24,
-      retry_on_exception=retry_api_check_quiet,
+      retry_on_exception=retry_util.retry_api_check_quiet,
       wait_exponential_multiplier=500,
       wait_exponential_max=64000)
   # For API errors dealing with auth, we want to retry, but not as often
   # Maximum 4 retries. Wait 1, 2, 4, 8 seconds.
   @retrying.retry(
       stop_max_attempt_number=5,
-      retry_on_exception=retry_auth_check_quiet,
+      retry_on_exception=retry_util.retry_auth_check_quiet,
       wait_exponential_multiplier=500,
       wait_exponential_max=8000)
   def execute_quiet(self, api):
